campo valor pago simbolo peso

+++ Intranet.web/Intranet.web/Models/NegociationViewModel.cs
@@ -23,10 +23,12 @@ namespace Intranet.web.Models
         public string Detalle { get; set; }
 
         [Required(ErrorMessage = "El campo {0} es requerido.")]
-        public int ValorNegociacion { get; set; }
+    
+        public long ValorNegociacion { get; set; }
 
         [Required(ErrorMessage = "El campo {0} es requerido.")]
-        public int BaseLiquidacion { get; set; }
+        
+        public long BaseLiquidacion { get; set; }
 
         [Required(ErrorMessage = "El campo {0} es requerido.")]
         [MaxLength(300)]
